Clean docs and streamline build
